requirements.txt
autogen
streamlit
streamlit-option-menu
streamlit_lottie
groq
scikit-learn
langchain
langchain-groq
streamlit-authenticator
bcrypt
pyrebase4
oauth2client
pycryptodome
gcloud
google-cloud-storage
setuptools
streamlit-lottie
requests==2.28.2
urllib3==1.26.15
requests-toolbelt==0.9.1
firebase_admin
python-dotenv
Pillow